requirements_ai.txt
# AI API Connections - Required Python Packages
# Install with: pip install -r requirements_ai.txt

# OpenAI Python SDK (works for both OpenAI and xAI/Grok)
openai>=1.12.0

# Environment variable management
python-dotenv>=1.0.0
